+++ CHANGELOG.md
@@ -12,8 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Removed display of dao url and dao ens name from `HeaderDao`, substituted with dropdown component
 -   Introduced required `daoAddress` in `HeaderDao`
 -   Made `daoEnsName` optional in `HeaderDao`
--   Renamed `copiedOnClick` to `copy` prop on `HeaderDao`, made `copy` accept copy input to be more generic
--   `shortenDaoUrl` util
+-   Renamed `copiedOnClick` to `onCopy` prop on `HeaderDao`, made `onCopy` accept copy input to be more generic
+-   Added `shortenDaoUrl` util
 
 ## [0.2.17] - 2023-09-26